[bug/Makefile]
Makefile 대규모 수정, syntax, tokenizer, wildcard, expand > src 디렉토리로 이동
