+++ include/minishell.h
@@ -3,16 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mingkim <mingkim@student.42.fr>            +#+  +:+       +#+        */
+/*   By: gimmingyu <gimmingyu@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 11:08:00 by mingkim           #+#    #+#             */
-/*   Updated: 2022/11/18 17:17:44 by mingkim          ###   ########.fr       */
+/*   Updated: 2022/11/19 00:28:21 by gimmingyu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
+# include "./base.h"
+# include "./struct.h"
 # include "./builtin.h"
 # include "./utils.h"
 # include "./core.h"
@@ -20,8 +22,6 @@
 # include "./executor.h"
 # include "./hashtable.h"
 # include "./tokenizer.h"
-# include "./struct.h"
-# include "./base.h"
 # include "./error.h"
 # include "./global.h"
 # include "./syntax.h"
