+++ src/internal/bintree/insert.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   insert.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mingkim <mingkim@student.42.fr>            +#+  +:+       +#+        */
+/*   By: gimmingyu <gimmingyu@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 15:12:19 by bolee             #+#    #+#             */
-/*   Updated: 2022/11/17 17:08:21 by mingkim          ###   ########.fr       */
+/*   Updated: 2022/11/19 00:58:37 by gimmingyu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ int	insert_rc_node(t_bintree_node *parent, t_token *token, t_tnType type)
 
 	if (parent == NULL || token == NULL)
 		return (FALSE);
-	node = create_bintree_node(token. type);
+	node = create_bintree_node(token, type);
 	if (node == NULL)
 		return (FALSE);
 	parent->rc = node;
