+++ src/internal/executor/bracket.c
