+++ src/internal/executor/double_quotes.c
