+++ src/internal/executor/executor.c
@@ -1,0 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   executor.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gimmingyu <gimmingyu@student.42.fr>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/10/14 15:24:57 by mingkim           #+#    #+#             */
+/*   Updated: 2022/11/19 01:18:11 by gimmingyu        ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../../include/minishell.h"
+
+int	executor(char *cmd)
+{
+	return (run_execute(cmd));
+}
+
+int	run_execute(char *cmd)
+{
+	int	status;
+
+	status = serve_status(cmd);
+	return (status);
+}
+
+int	serve_status(char *cmd)
+{
+	int	status;
+
+	printf("cmd = %s\n", cmd);
+	status = EXIT_SUCCESS;
+	return (status);
+}
