+++ src/internal/executor/pipe.c
