+++ src/internal/tokenizer/verify_brackets.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   verify_brackets.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mingkim <mingkim@student.42.fr>            +#+  +:+       +#+        */
+/*   By: gimmingyu <gimmingyu@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/18 16:05:42 by mingkim           #+#    #+#             */
-/*   Updated: 2022/11/18 17:13:55 by mingkim          ###   ########.fr       */
+/*   Updated: 2022/11/19 00:38:03 by gimmingyu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../../include/minishell.h"
+#include "../../../include/minishell.h"
 
 t_tType	verify_bracket(char *str)
 {
